+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## main
 
+* refactor - rename parameter `skipFirst*` to `skipInitial*`
 
 ## [0.1.0-beta.5] - 2020-08-06