+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## main
 
 * refactor - rename parameter `skipFirst*` to `skipInitial*`
+* refactor - reimplement `system.react` and `system.reactLatest` operators
 
 ## [0.1.0-beta.5] - 2020-08-06