+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * refactor - rename parameter `skipFirst*` to `skipInitial*`
 * refactor - reimplement `system.react` and `system.reactLatest` operators
+* feature - add `system.log` operator
 
 ## [0.1.0-beta.5] - 2020-08-06