+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * refactor - rename parameter `skipFirst*` to `skipInitial*`
 * refactor - reimplement `system.react` and `system.reactLatest` operators
 * feature - add `system.log` operator
+* feature - add `system.eventInterceptor` operator to intercept event
 
 ## [0.1.0-beta.5] - 2020-08-06